Fixed sequential simulation runs hanging issue
The UDP socket created in one simulation is closed only when the second
simulation is started (not when the first is closed). This, however, does
not happen on windows, instead the first socket remains open and all packets
get routed to that instead of to the second one (causing the second
simulation to hang). To fix this issue:

- Added close method for UDP socket processes
- Modified FpgaPesEnsembleNetwork to call UDP socket process close method
  when the simulator object is closed